Finished Version

+++ Nim+/Program.cs
@@ -220,7 +220,7 @@ Console.WriteLine("Lets Play !");
                 {
                     drawPlayer2I = remainingMatches - 1;
                 }
-                drawPlayer2I = (int)Math.Clamp((double)drawPlayer2I, 1, 3);
+                drawPlayer2I = (int)Math.Clamp(drawPlayer2I, 1, 3);
                 Console.Write("HAL2000 draws ");
                 Console.WriteLine(drawPlayer2I);
